# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AuthFramework REST API
  description: |
    Comprehensive REST API for the AuthFramework authentication and authorization library.

    This API provides complete access to AuthFramework's authentication, authorization,
    user management, multi-factor authentication, OAuth 2.0, and administrative features
    through standardized HTTP endpoints.

    ## Authentication

    Most endpoints require authentication via Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    ## Rate Limiting

    API endpoints are rate-limited to prevent abuse:
    - Authentication endpoints: 5 requests per minute per IP
    - Standard endpoints: 100 requests per minute per user
    - Admin endpoints: 50 requests per minute per admin user

    ## Error Handling

    All endpoints return standardized error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "INVALID_CREDENTIALS",
        "message": "The provided credentials are invalid",
        "details": null
      },
      "timestamp": "2024-01-20T15:30:45Z"
    }
    ```
  version: 1.0.0
  contact:
    name: AuthFramework Support
    url: https://github.com/cires/AuthFramework
    email: support@authframework.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.authframework.dev
    description: Production server
  - url: https://staging-api.authframework.dev
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return access and refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimit"
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get current user's profile information
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags:
        - Users
      summary: Update user profile
      description: Update current user's profile information
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when rate limit resets
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response payload (varies by endpoint)
        timestamp:
          type: string
          format: date-time
          example: "2024-01-20T15:30:45Z"
      required:
        - success
        - timestamp
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_CREDENTIALS"
            message:
              type: string
              example: "The provided credentials are invalid"
            details:
              type: object
              nullable: true
          required:
            - code
            - message
        timestamp:
          type: string
          format: date-time
          example: "2024-01-20T15:30:45Z"
      required:
        - success
        - error
        - timestamp
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePassword123!"
        remember_me:
          type: boolean
          default: false
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 3600
        user:
          $ref: "#/components/schemas/UserInfo"
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - user
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "user@example.com"
        email:
          type: string
          format: email
          example: "user@example.com"
        roles:
          type: array
          items:
            type: string
          example: ["user"]
        mfa_enabled:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2024-01-20T15:30:45Z"
        last_login:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-20T15:30:45Z"
      required:
        - id
        - username
        - email
        - roles
        - mfa_enabled
        - created_at
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "user@example.com"
        email:
          type: string
          format: email
          example: "user@example.com"
        first_name:
          type: string
          nullable: true
          example: "John"
        last_name:
          type: string
          nullable: true
          example: "Doe"
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        timezone:
          type: string
          nullable: true
          example: "UTC"
        locale:
          type: string
          nullable: true
          example: "en-US"
        mfa_enabled:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2024-01-20T15:30:45Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-20T15:30:45Z"
      required:
        - id
        - username
        - email
        - mfa_enabled
        - created_at
        - updated_at
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
          example: "John"
        last_name:
          type: string
          nullable: true
          example: "Doe"
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        timezone:
          type: string
          nullable: true
          example: "UTC"
        locale:
          type: string
          nullable: true
          example: "en-US"
...